通过最小公倍数，将浮点运算换成两个整形来计算，成功实现优化的高响应比调度算法。

+++ src/kernel/main.c
@@ -214,8 +214,8 @@ PRIVATE void init_origin() {
 
     origin->level = 5;
 #ifdef LEVEL_SCHEDULE
-    origin->wait = 0.0;
-    origin->service = (float) origin->level;
+    origin->wait = 0;
+    origin->service = origin->level;
 #endif
     ready(origin);
 }
